backend/requirements.txt
django==5.2.7
djangorestframework==3.16.1
graphene-django==3.2.3
pymongo==4.15.2
django-cors-headers==4.9.0
python-decouple==3.8
